--- src/main/java/roombook/appointment/AvailabilityBlock.java
@@ -1,7 +1,0 @@
-package roombook.appointment;
-
-/**
- * Created by rgingeleski on 3/5/2018.
- */
-public class AvailabilityBlock {
-}
