+++ src/main/java/roombook/model/appointment/AvailabilityBlock.java
@@ -1,0 +1,7 @@
+package roombook.model.appointment;
+
+/**
+ * Created by rgingeleski on 3/5/2018.
+ */
+public class AvailabilityBlock {
+}
