+++ src/main/java/roombook/model/room/ConferenceRoom.java
@@ -1,0 +1,9 @@
+package roombook.model.room;
+
+public class ConferenceRoom extends Room
+{
+    public String getType()
+    {
+        return "CONFERENCE";
+    }
+}
