+++ src/main/java/roombook/model/room/FocusRoom.java
@@ -1,0 +1,9 @@
+package roombook.model.room;
+
+public class FocusRoom extends Room
+{
+    public String getType()
+    {
+        return "FOCUS";
+    }
+}
