+++ src/main/java/roombook/model/room/ShareRoom.java
@@ -1,0 +1,9 @@
+package roombook.model.room;
+
+public class ShareRoom extends Room
+{
+    public String getType()
+    {
+        return "SHARE";
+    }
+}
