+++ src/main/java/roombook/model/user/AdminUser.java
@@ -1,0 +1,9 @@
+package roombook.model.user;
+
+public class AdminUser extends User
+{
+    public String getType()
+    {
+        return "ADMIN";
+    }
+}
