+++ src/main/java/roombook/model/user/AssistantUser.java
@@ -1,0 +1,9 @@
+package roombook.model.user;
+
+public class AssistantUser extends User
+{
+    public String getType()
+    {
+        return "ASSISTANT";
+    }
+}
