+++ src/main/java/roombook/model/user/EmployeeUser.java
@@ -1,0 +1,9 @@
+package roombook.model.user;
+
+public class EmployeeUser extends User
+{
+    public String getType()
+    {
+        return "EMPLOYEE";
+    }
+}
