+++ src/main/java/roombook/model/user/HRManagerUser.java
@@ -1,0 +1,9 @@
+package roombook.model.user;
+
+public class HRManagerUser extends User
+{
+    public String getType()
+    {
+        return "HR_MGR";
+    }
+}
