--- src/main/java/roombook/room/ConferenceRoom.java
@@ -1,9 +1,0 @@
-package roombook.room;
-
-public class ConferenceRoom extends Room
-{
-    public String getType()
-    {
-        return "CONFERENCE";
-    }
-}
