--- src/main/java/roombook/room/FocusRoom.java
@@ -1,9 +1,0 @@
-package roombook.room;
-
-public class FocusRoom extends Room
-{
-    public String getType()
-    {
-        return "FOCUS";
-    }
-}
