--- src/main/java/roombook/room/ShareRoom.java
@@ -1,9 +1,0 @@
-package roombook.room;
-
-public class ShareRoom extends Room
-{
-    public String getType()
-    {
-        return "SHARE";
-    }
-}
