--- src/main/java/roombook/user/AdminUser.java
@@ -1,9 +1,0 @@
-package roombook.user;
-
-public class AdminUser extends User
-{
-    public String getType()
-    {
-        return "ADMIN";
-    }
-}
