--- src/main/java/roombook/user/AssistantUser.java
@@ -1,9 +1,0 @@
-package roombook.user;
-
-public class AssistantUser extends User
-{
-    public String getType()
-    {
-        return "ASSISTANT";
-    }
-}
