--- src/main/java/roombook/user/EmployeeUser.java
@@ -1,9 +1,0 @@
-package roombook.user;
-
-public class EmployeeUser extends User
-{
-    public String getType()
-    {
-        return "EMPLOYEE";
-    }
-}
