--- src/main/java/roombook/user/HRManagerUser.java
@@ -1,9 +1,0 @@
-package roombook.user;
-
-public class HRManagerUser extends User
-{
-    public String getType()
-    {
-        return "HR_MGR";
-    }
-}
