--- src/main/java/roombook/user/User.java
@@ -1,51 +1,0 @@
-package roombook.user;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-public abstract class User
-{
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    private String username;
-    private String password;
-
-    public User()
-    {
-        this.setUsername(null);
-        this.setPassword(null);
-    }
-
-    public User(String name, String username, String password)
-    {
-        this.setUsername(username);
-        this.setPassword(password);
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public abstract String getType();
-}
